requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation
pydantic==2.5.0

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Async file operations
aiofiles==23.2.1
